move multidim module
