+++ etl/collections/__init__.py
@@ -1,0 +1,5 @@
+"""This module is a package for collections of indicators. This includes:
+
+- Explorers
+- MDIMs
+"""
