+++ etl/steps/export/multidim/health/latest/causes_of_death.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from etl import multidim
+from etl.collections import multidim
 from etl.helpers import PathFinder
 
 # Get paths and naming conventions for current step.
